+++ packages/services/pkg/utils/general.go
@@ -1,0 +1,8 @@
+package utils
+
+func Min(a, b int) int {
+	if a < b {
+		return a
+	}
+	return b
+}